+++ config.py
@@ -11,7 +11,7 @@ Auteur        : Pierre Théberge
 Compagnie     : Innovations, Performances, Technologies inc.
 Créé le       : 2025-08-05
 Modifié le    : 2026-08-18
-Version       : 0.5.20
+Version       : 0.5.21
 Copyright     : Pierre Théberge
 
 Description
@@ -111,6 +111,7 @@ Modifications
 0.5.18 - 2026-07-10   [CR]    : Synchronisation de version (aucun changement fonctionnel).
 0.5.19 - 2026-08-15   [CR]    : Synchronisation de version (aucun changement fonctionnel).
 0.5.20 - 2026-08-18   ES-34   : Synchronisation de version (aucun changement fonctionnel).
+0.5.21 - 2026-08-18   ES-34   : Synchronisation de version (aucun changement fonctionnel).
 
 Paramètres
 ----------